requirements.txt
python-telegram-bot==21.0.1
openai==1.3.6
python-dotenv==1.0.1
aiohttp==3.9.1